+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [2.11.3]
+- Adds methods `enableGestures` and `enableKeyboard` in `gameRef`(BonfireGameInterface)
+- Adds mixin `KeyboardEventListener`.
+- Fix exception in `TiledWorldBuilder`.
+
 # [2.11.2]
 - BugFix[`BarLifeComponent`]: animate in web.